% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{n2CrossKnn}
\alias{n2CrossKnn}
\title{find NN of A in B}
\usage{
n2CrossKnn(
  mA,
  mB,
  k,
  nThreads = 10L,
  verbose = TRUE,
  indexType = "angular",
  M = 12L,
  MaxM0 = 24L,
  ef_search_multiplier = 50,
  quiet = FALSE
)
}
\arguments{
\item{mA}{matrix i.e. Rcpp::NumericMatrix}

\item{mB}{matrix i.e. Rcpp::NumericMatrix}

\item{k}{int}

\item{nThreads}{integer (default=10)}

\item{verbose}{boolean (default=FALSE)}

\item{indexType}{string (default="angular")}

\item{M}{integer (default=12)}

\item{MaxM0}{integer (default=24)}

\item{ef_search_multiplier}{numeric (default=50)}

\item{quiet}{boolean (default=FALSE)}
}
\description{
find NN of A in B
}
